Extract fetchBlocks into its own hook, Add blockexplorer to the index page.
