+++ packages/nextjs/hooks/scaffold-eth/index.ts
@@ -12,3 +12,4 @@ export * from "./useScaffoldContractWrite";
 export * from "./useScaffoldEventSubscriber";
 export * from "./useScaffoldEventHistory";
 export * from "./useTransactor";
+export * from "./useFetchBlocks";
